% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/covariance.R
\name{assume_iid}
\alias{assume_iid}
\title{Produces an identity matrix}
\usage{
assume_iid(X, Y, B, filt = rep(TRUE, dim(Y)[2]), ...)
}
\arguments{
\item{X}{design matrix}

\item{Y}{response matrix}

\item{B}{fitted coefficient matrix}

\item{filt}{Logical vector: which responses to use}

\item{shrink}{shrinkage factor}

\item{mc.cores}{parallelization}
}
\value{
Estimate for covariance matrix
}
\description{
A \code{Sigma_t_method}.
}

