% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bayes.R
\name{fast_ridge}
\alias{fast_ridge}
\title{Ridge regression subroutine}
\usage{
fast_ridge(X, V_X, d_X, lambda, Y, naive = FALSE)
}
\arguments{
\item{X}{covariate/design matrix}

\item{V_X}{from svd of X}

\item{d_X}{from svd of X}

\item{lambda}{ridge regularization}

\item{Y}{response matrix}

\item{naive}{Do the naive computation (for testing)}
}
\value{
a matrix B
}
\description{
computes solve(xtx + lambda, t(X) %*% Y)
as 1/lambda * t(X) %*% Y
V_X, d_X are from svd of X
}

