% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forward.R
\name{fit_ridge_kernel}
\alias{fit_ridge_kernel}
\title{Multivariate ridge regression}
\usage{
fit_ridge_kernel(X, Y, lambda = 1, mc.cores = 3, ...)
}
\arguments{
\item{X}{design matrix}

\item{Y}{response matrix}

\item{lambda}{L2 regularization}

\item{mc.cores}{number of mc cores}
}
\value{
B coefficient matrix
}
\description{
Fits multivariate ridge with the same lambda for all columns
Computes \code{B = solve(t(X)\%*\% X + lambda * eye(p), t(X) \%*\% Y)}.
Uses kernel trick and parallelization.
}

