% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bayes.R
\name{kron_v}
\alias{kron_v}
\title{Computes \code{(A \%x\% B) \%*\% v}}
\usage{
kron_v(A, B, cc)
}
\arguments{
\item{A}{matrix of size a x b}

\item{B}{matrix of size c x d}

\item{cc}{vector of length b * d}
}
\value{
vector of size a * c
}
\description{
Computes \code{(A \%x\% B) \%*\% v}
}

