% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bayes.R
\name{post_moments}
\alias{post_moments}
\title{Computes posterior moments for Bayesian multivariate regression}
\usage{
post_moments(X, Y, Sigma_e, Sigma_b, Sigma_t = eye(dim(X)[1]),
  computeCov = TRUE, naive = FALSE, matrix = TRUE, ...)
}
\arguments{
\item{X}{covariate matrix}

\item{Y}{response matrix}

\item{Sigma_e}{Covariance within given time (row) of Y}

\item{Sigma_b}{Prior covariance of a column of B}

\item{Sigma_t}{Autocorrelation of a single response (column) of Y over time}

\item{computeCov}{Return the posterior covariance of B?}

\item{naive}{Do the naive computation (for testing)}

\item{matrix}{Return B as a matrix? (or vector)}
}
\value{
a matrix/vector (if \code{computeCov = FALSE})
or a list (if \code{computeCov = TRUE})
}
\description{
Posterior mean and covariance given a prior on Bvec
 of the form Sigma_b %x% eye(pX)
 and noise prior of form Sigma_e %x% Sigma_t
}

