% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/backward.R
\name{pre_mle}
\alias{pre_mle}
\title{Predictive distribution corresponding to MLE rule}
\usage{
pre_mle(X_te, B, Sigma_e, filt, ...)
}
\arguments{
\item{X_te}{Test class covariates}

\item{B}{fitted coefficients}

\item{Sigma_e}{Fitted response covariance}

\item{filt}{Logical vector: which responses to use}
}
\value{
A list with moments
}
\description{
A \code{backward_method}.
MLE rule is equivalent to using the same covariance for each response
}

