% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bayes.R
\name{tkron_d_kron}
\alias{tkron_d_kron}
\title{Computes t(A %x% B) %*% diag(d) %*% (A %*% B)}
\usage{
tkron_d_kron(A, B, d)
}
\arguments{
\item{A}{matrix of size a x b}

\item{B}{matrix of size c x d}

\item{d}{vector: represents diagonal matrix of size a * c}
}
\value{
square matrix of size a * c
}
\description{
Computes t(A %x% B) %*% diag(d) %*% (A %*% B)
}

